Use pipelining mode
So pipelining mode essentially means to send a couple of ansible
requests at once rather than a connection per task. This massively
speeds up deployment for me rather than have SSH renegotiate each time.

The down side of this is that you have to have tty-less sudo enabled
first. This is not how rhel machines ship by default. Enabling
pipelining is also something that you do as a global config option not
per playbook.

To get around this use a paramiko connection (which doesn't work with
pipelining) to turn off the tty requirement of the VM then run the rest
of the playbook using SSH pipelining enabled.

I had to set host_key_checking=False to allow this to work because it
appears paramiko uses a different method of host key acceptance than SSH
so it wasn't picking up known machines.